requirements.txt
selenium==3.141.0
Pillow==8.3.1
numpy